Move db_gateways.py to /nanny folder.
