+++ login_app/views/existing_user_sign_in.py
@@ -8,7 +8,7 @@ from nanny import notify
 from nanny import utilities
 from .base import BaseFormView
 
-from db_gateways import IdentityGatewayActions
+from nanny.db_gateways import IdentityGatewayActions
 
 
 class ExistingUserSignInFormView(BaseFormView):
